requirements.txt
annotated-types==0.7.0
anyio==4.9.0
# Editable Git install with no remote (ata-backend==0.1.0)
-e c:\users\user\desktop\ata demo\ata-project\ata-backend
backports.asyncio.runner==1.2.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fonttools==4.59.0
fpdf2==2.8.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
RestrictedPython==8.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
